Bug Fixing
Fixed errors in the order of recursive calls that was allowing
statements such as "+ A + B" to be accepted.

Also fixed the way error and accept were written to be actual functions
rather than simply returning a string. This negated the need for the
call to println before the trampoline.